# Use `slices.Sort()` instead of `sort.Strings()` or `sort.Ints()`