--- a/Makefile
+++ b/Makefile
@@ -316,6 +316,15 @@ lint: check-makefiles
 		github.com/prometheus/client_golang/prometheus.{NewCounter,NewCounterVec,NewCounterFunc,NewGauge,NewGaugeVec,NewGaugeFunc,NewSummary,NewSummaryVec,NewHistogram,NewHistogramVec}=github.com/prometheus/client_golang/prometheus/promauto.With" \
 		./pkg/...
 
+	# Use the faster slices.Sort where we can.
+	# See https://github.com/prometheus/prometheus/pull/11054 and https://github.com/prometheus/prometheus/pull/11318 for more
+	# background information on slices.Sort() and the real-world performance impact it can have.
+	# Note that we don't automatically suggest replacing sort.Float64s() with slices.Sort() as the documentation for slices.Sort()
+	# at the time of writing warns that slices.Sort() may not correctly handle NaN values.
+	faillint -paths \
+		"sort.{Strings,Ints}=golang.org/x/exp/slices.Sort" \
+		./pkg/... ./cmd/... ./tools/... ./integration/...
+
 format: ## Run gofmt and goimports.
 	find . $(DONT_FIND) -name '*.pb.go' -prune -o -type f -name '*.go' -exec gofmt -w -s {} \;
 	find . $(DONT_FIND) -name '*.pb.go' -prune -o -type f -name '*.go' -exec goimports -w -local github.com/grafana/mimir {} \;