--- a/pkg/distributor/distributor_test.go
+++ b/pkg/distributor/distributor_test.go
@@ -39,6 +39,7 @@ import (
 	"github.com/weaveworks/common/mtime"
 	"github.com/weaveworks/common/user"
 	"go.uber.org/atomic"
+	"golang.org/x/exp/slices"
 	"golang.org/x/time/rate"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/health/grpc_health_v1"
@@ -2254,7 +2255,7 @@ func TestDistributor_LabelNamesAndValues(t *testing.T) {
 
 			// sort label values to make stable assertion
 			for _, item := range response.Items {
-				sort.Strings(item.Values)
+				slices.Sort(item.Values)
 			}
 			assert.ElementsMatch(t, response.Items, expectedLabelValues)
 		})
@@ -3998,7 +3999,7 @@ func (i *mockIngester) LabelNames(ctx context.Context, req *client.LabelNamesReq
 			}
 		}
 	}
-	sort.Strings(response.LabelNames)
+	slices.Sort(response.LabelNames)
 
 	return &response, nil
 }