--- a/pkg/scheduler/queue/user_queues.go
+++ b/pkg/scheduler/queue/user_queues.go
@@ -10,6 +10,8 @@ import (
 	"sort"
 	"time"
 
+	"golang.org/x/exp/slices"
+
 	"github.com/grafana/mimir/pkg/util"
 )
 
@@ -197,7 +199,7 @@ func (q *queues) addQuerierConnection(querierID string) {
 	// First connection from this querier.
 	q.queriers[querierID] = &querier{connections: 1}
 	q.sortedQueriers = append(q.sortedQueriers, querierID)
-	sort.Strings(q.sortedQueriers)
+	slices.Sort(q.sortedQueriers)
 
 	q.recomputeUserQueriers()
 }