--- a/pkg/storegateway/indexheader/reader_benchmarks_test.go
+++ b/pkg/storegateway/indexheader/reader_benchmarks_test.go
@@ -20,6 +20,7 @@ import (
 	"github.com/grafana/mimir/pkg/storage/tsdb/block"
 	"github.com/grafana/mimir/pkg/storage/tsdb/metadata"
 	"github.com/grafana/mimir/pkg/storegateway/testhelper"
+	"github.com/grafana/mimir/pkg/util/test"
 )
 
 func BenchmarkLookupSymbol(b *testing.B) {
@@ -138,6 +139,55 @@ func BenchmarkLabelNames(b *testing.B) {
 			})
 		}
 	}
+}
+
+func BenchmarkLabelValuesIndexV1(b *testing.B) {
+	ctx := context.Background()
+
+	bucketDir := b.TempDir()
+	bkt, err := filesystem.NewBucket(filepath.Join(bucketDir, "bkt"))
+	require.NoError(b, err)
+	b.Cleanup(func() {
+		require.NoError(b, bkt.Close())
+	})
+
+	metaIndexV1, err := metadata.ReadFromDir("./testdata/index_format_v1")
+	require.NoError(b, err)
+	test.Copy(b, "./testdata/index_format_v1", filepath.Join(bucketDir, metaIndexV1.ULID.String()))
+
+	_, err = metadata.InjectThanos(log.NewNopLogger(), filepath.Join(bucketDir, metaIndexV1.ULID.String()), metadata.Thanos{
+		Labels:     labels.FromStrings("ext1", "1").Map(),
+		Downsample: metadata.ThanosDownsample{Resolution: 0},
+		Source:     metadata.TestSource,
+	}, &metaIndexV1.BlockMeta)
+
+	require.NoError(b, err)
+	require.NoError(b, block.Upload(ctx, log.NewNopLogger(), bkt, filepath.Join(bucketDir, metaIndexV1.ULID.String()), nil))
+
+	indexName := filepath.Join(bucketDir, metaIndexV1.ULID.String(), block.IndexHeaderFilename)
+	require.NoError(b, WriteBinary(ctx, bkt, metaIndexV1.ULID, indexName))
+
+	benchmarkReaders(b, bucketDir, metaIndexV1.ULID, func(b *testing.B, br Reader) {
+		names, err := br.LabelNames()
+		require.NoError(b, err)
+
+		rand.Shuffle(len(names), func(i, j int) {
+			names[i], names[j] = names[j], names[i]
+		})
+
+		b.ResetTimer()
+
+		for i := 0; i < b.N; i++ {
+			name := names[i%len(names)]
+
+			values, err := br.LabelValues(name, func(s string) bool {
+				return true
+			})
+
+			require.NoError(b, err)
+			require.NotEmpty(b, values)
+		}
+	})
 }
 
 func benchmarkReaders(b *testing.B, bucketDir string, id ulid.ULID, benchmark func(b *testing.B, br Reader)) {