#### What this PR does

This PR was sparked by https://github.com/grafana/mimir/pull/3639#discussion_r1042880752. It replaces all usage of `sort.Strings()` with [`slices.Sort()`](https://pkg.go.dev/golang.org/x/exp@v0.0.0-20221211140036-ad323defaf05/slices#Sort), and adds a linting rule to enforce this in the future.

I'm opening this as a draft for discussion because:
* this is a wide-ranging change
* `golang.org/x/exp/slices` is still considered an experimental package (although Prometheus is now using it [widely](https://github.com/prometheus/prometheus/pull/11318))
* I haven't benchmarked these changes (but everywhere else that has replaced `sort.Strings()` with `slices.Sort()` has seen substantial improvements, see below)

---

##### Performance

https://github.com/grafana/mimir/pull/3639#discussion_r1042880752 and https://github.com/prometheus/prometheus/pull/11054 show substantial performance improvements when using `slices.Sort()` over `sort.Strings()`. This improved performance is because `slices.Sort()` does not require the overhead of an interface call for every comparison. (If we were using Go 1.18 or earlier, we'd also benefit from the improved pdquicksort sorting algorithm in `slices`, but this has been implemented in the `sort` package as of Go 1.19.)

Given the pattern of significant performance improvements in a number of scenarios, I haven't benchmarked the changes in this PR. If we do want to take a more measured approach, I'm very open to that feedback.

##### Future work

We could also replace `sort.Float64s()` with `slices.Sort()`, however, this requires more careful consideration as [the documentation for `slices.Sort()`](https://pkg.go.dev/golang.org/x/exp@v0.0.0-20221211140036-ad323defaf05/slices#Sort) warns that it may not handle NaN values correctly.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [n/a] Tests updated
- [n/a] Documentation added
- [n/a] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`