@bboreham 2022-12-12
> This improved performance is because slices.Sort() does not require the overhead of an interface call for every comparison.

This is partially true; as I said in the Prometheus PR.

> Some of the speedup comes from [not] an `interface .Less()` call; some comes from `exp/slices` using "pattern-defeating quicksort(pdqsort)" which is much better on already-sorted data.