@charleskorn 2022-12-13
> This is partially true; as I said in the Prometheus PR.
> 
> > Some of the speedup comes from [not] an `interface .Less()` call; some comes from `exp/slices` using "pattern-defeating quicksort(pdqsort)" which is much better on already-sorted data.

I think this changed with the Go 1.19 release - [the release notes](https://go.dev/doc/go1.19) mention changing the algorithm for the `sort` package to pdqsort, and `sort.Strings()` calls into [an implementation of pdqsort](https://cs.opensource.google/go/go/+/refs/tags/go1.19.4:src/sort/zsortinterface.go;drc=888047c31007c0ad96764254fa8afb1417ddfd40;l=61) which looks almost identical to [the one in `golang.org/x/exp/slices`](https://cs.opensource.google/go/x/exp/+/ad323def:slices/zsortordered.go;drc=732eee02a75a571a8347fdf60af8c8b9e3ca9254;l=63).