@pracucci 2022-12-13
@charleskorn I would keep a very practical approach, benchmarking few of the changes in the hot path, but not all of them. If we don't find regressions on few of them, I would just move on and merge.

I think this PR touches somewhat hot paths in the following functions. Do we have any benchmark touching any of those?

```
Distributor.LabelValuesForLabelName
Distributor.LabelNames
mergeExemplarQueryResponses()
matrixMerge()
fetchLabelValuesFromStore()
blockLabelNames()
BinaryReader.LabelValues()
BinaryReader.LabelNames()
PostingOffsetTableV1.LabelValues()
```