@charleskorn 2022-12-14
tl;dr: for the four instances I benchmarked, we see improvements of anywhere from 2.2% to 16%, and none regressed. I'm happy to keep benchmarking the other changes, but the results are universally positive so far.

---

These are some results from benchmarks that we already had or were easily able to be changed to cover some of the affected methods:

<details>
<summary><code>BinaryReader.LabelNames()</code></summary>

```
name                                                 old time/op    new time/op    delta
LabelNames/20Names100Values/BinaryReader-10            1.69µs ± 2%    1.48µs ± 2%  -12.21%  (p=0.008 n=5+5)
LabelNames/20Names500Values/BinaryReader-10            1.69µs ± 1%    1.49µs ± 2%  -11.98%  (p=0.008 n=5+5)
LabelNames/20Names1000Values/BinaryReader-10           1.70µs ± 2%    1.49µs ± 1%  -12.25%  (p=0.008 n=5+5)
LabelNames/50Names100Values/BinaryReader-10            4.40µs ± 1%    3.73µs ± 1%  -15.06%  (p=0.008 n=5+5)
LabelNames/50Names500Values/BinaryReader-10            4.46µs ± 5%    3.73µs ± 1%  -16.45%  (p=0.008 n=5+5)
LabelNames/50Names1000Values/BinaryReader-10           4.37µs ± 1%    3.72µs ± 1%  -14.75%  (p=0.008 n=5+5)
LabelNames/100Names100Values/BinaryReader-10           9.51µs ± 0%    8.18µs ± 0%  -14.00%  (p=0.008 n=5+5)
LabelNames/100Names500Values/BinaryReader-10           9.53µs ± 1%    8.19µs ± 1%  -14.05%  (p=0.008 n=5+5)
LabelNames/100Names1000Values/BinaryReader-10          9.56µs ± 1%    8.18µs ± 1%  -14.45%  (p=0.008 n=5+5)
LabelNames/200Names100Values/BinaryReader-10           20.6µs ± 0%    17.5µs ± 0%  -15.00%  (p=0.008 n=5+5)
LabelNames/200Names500Values/BinaryReader-10           20.6µs ± 0%    17.5µs ± 0%  -15.09%  (p=0.008 n=5+5)
LabelNames/200Names1000Values/BinaryReader-10          20.8µs ± 3%    17.4µs ± 1%  -16.08%  (p=0.008 n=5+5)

name                                                 old alloc/op   new alloc/op   delta
LabelNames/20Names100Values/BinaryReader-10              424B ± 0%      400B ± 0%   -5.66%  (p=0.008 n=5+5)
LabelNames/20Names500Values/BinaryReader-10              424B ± 0%      400B ± 0%   -5.66%  (p=0.008 n=5+5)
LabelNames/20Names1000Values/BinaryReader-10             424B ± 0%      400B ± 0%   -5.66%  (p=0.008 n=5+5)
LabelNames/50Names100Values/BinaryReader-10              968B ± 0%      944B ± 0%   -2.48%  (p=0.008 n=5+5)
LabelNames/50Names500Values/BinaryReader-10              968B ± 0%      944B ± 0%   -2.48%  (p=0.008 n=5+5)
LabelNames/50Names1000Values/BinaryReader-10             968B ± 0%      944B ± 0%   -2.48%  (p=0.008 n=5+5)
LabelNames/100Names100Values/BinaryReader-10           1.86kB ± 0%    1.84kB ± 0%   -1.29%  (p=0.008 n=5+5)
LabelNames/100Names500Values/BinaryReader-10           1.86kB ± 0%    1.84kB ± 0%   -1.29%  (p=0.008 n=5+5)
LabelNames/100Names1000Values/BinaryReader-10          1.86kB ± 0%    1.84kB ± 0%   -1.29%  (p=0.008 n=5+5)
LabelNames/200Names100Values/BinaryReader-10           3.53kB ± 0%    3.50kB ± 0%   -0.68%  (p=0.008 n=5+5)
LabelNames/200Names500Values/BinaryReader-10           3.53kB ± 0%    3.50kB ± 0%   -0.68%  (p=0.008 n=5+5)
LabelNames/200Names1000Values/BinaryReader-10          3.53kB ± 0%    3.50kB ± 0%   -0.68%  (p=0.008 n=5+5)

name                                                 old allocs/op  new allocs/op  delta
LabelNames/20Names100Values/BinaryReader-10              4.00 ± 0%      3.00 ± 0%  -25.00%  (p=0.008 n=5+5)
LabelNames/20Names500Values/BinaryReader-10              4.00 ± 0%      3.00 ± 0%  -25.00%  (p=0.008 n=5+5)
LabelNames/20Names1000Values/BinaryReader-10             4.00 ± 0%      3.00 ± 0%  -25.00%  (p=0.008 n=5+5)
LabelNames/50Names100Values/BinaryReader-10              4.00 ± 0%      3.00 ± 0%  -25.00%  (p=0.008 n=5+5)
LabelNames/50Names500Values/BinaryReader-10              4.00 ± 0%      3.00 ± 0%  -25.00%  (p=0.008 n=5+5)
LabelNames/50Names1000Values/BinaryReader-10             4.00 ± 0%      3.00 ± 0%  -25.00%  (p=0.008 n=5+5)
LabelNames/100Names100Values/BinaryReader-10             4.00 ± 0%      3.00 ± 0%  -25.00%  (p=0.008 n=5+5)
LabelNames/100Names500Values/BinaryReader-10             4.00 ± 0%      3.00 ± 0%  -25.00%  (p=0.008 n=5+5)
LabelNames/100Names1000Values/BinaryReader-10            4.00 ± 0%      3.00 ± 0%  -25.00%  (p=0.008 n=5+5)
LabelNames/200Names100Values/BinaryReader-10             4.00 ± 0%      3.00 ± 0%  -25.00%  (p=0.008 n=5+5)
LabelNames/200Names500Values/BinaryReader-10             4.00 ± 0%      3.00 ± 0%  -25.00%  (p=0.008 n=5+5)
LabelNames/200Names1000Values/BinaryReader-10            4.00 ± 0%      3.00 ± 0%  -25.00%  (p=0.008 n=5+5)
```

</details>

<details>
<summary><code>BinaryReader.LabelValues()</code> and <code>PostingOffsetTableV1.LabelValues()</code></summary>

Note: `StreamBinaryReader` uses `PostingOffsetTableV1` - I've tested it through `StreamBinaryReader` here to save me constructing a slightly different benchmark setup just for it.

```
name                                      old time/op    new time/op    delta
LabelValuesIndexV1/StreamBinaryReader-10    4.19µs ± 1%    3.55µs ± 2%  -15.23%  (p=0.008 n=5+5)
LabelValuesIndexV1/BinaryReader-10          4.18µs ± 1%    3.53µs ± 0%  -15.62%  (p=0.008 n=5+5)

name                                      old alloc/op   new alloc/op   delta
LabelValuesIndexV1/StreamBinaryReader-10      960B ± 0%      936B ± 0%   -2.50%  (p=0.008 n=5+5)
LabelValuesIndexV1/BinaryReader-10            960B ± 0%      936B ± 0%   -2.50%  (p=0.008 n=5+5)

name                                      old allocs/op  new allocs/op  delta
LabelValuesIndexV1/StreamBinaryReader-10      3.00 ± 0%      2.00 ± 0%  -33.33%  (p=0.008 n=5+5)
LabelValuesIndexV1/BinaryReader-10            3.00 ± 0%      2.00 ± 0%  -33.33%  (p=0.008 n=5+5)
```

</details>

<details>
<summary><code>mergeExemplarQueryResponses</code></summary>

```
name               old time/op    new time/op    delta
MergeExemplars-10    1.15ms ± 0%    1.13ms ± 0%  -2.20%  (p=0.008 n=5+5)

name               old alloc/op   new alloc/op   delta
MergeExemplars-10    1.28MB ± 0%    1.28MB ± 0%    ~     (p=0.794 n=5+5)

name               old allocs/op  new allocs/op  delta
MergeExemplars-10     5.06k ± 0%     5.06k ± 0%  -0.02%  (p=0.008 n=5+5)
```

</details>